+++ src/features/Users/components/RecentUsers.tsx
@@ -3,7 +3,7 @@ import { columns } from '../columns';
 import { useUsers } from '../hooks/useUsers';
 import TableView from './TableView';
 import { DEFAULT_LIMIT } from '../constants';
-import { TableProvider } from '../context/table.context';
+import { TableProvider } from '../context/users-table.context';
 
 const DEFAULT_PROPS = { page: 0, limit: DEFAULT_LIMIT };
